refactor: move useTableExpand to VsTableBodyRow
